Timers
These include the timers so that the piston doesnt break when the motors start running. This DOES NOT INCLUDE the timer used when the robot is interuppted, shut off, or ends a match. We need to get the timer from the field of the match so we shut down motors and wait half a secto put piston in so we stay where we are no matter what.